# docs/swagger.yaml
basePath: /
definitions:
  model.CreateSubscriptionRequest:
    properties:
      cost_rub:
        type: integer
      end_date:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - cost_rub
    - service_name
    - start_date
    - user_id
    type: object
  model.Subscription:
    properties:
      cost_rub:
        type: integer
      end_date:
        type: string
      id:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    type: object
  model.UpdateSubscriptionRequest:
    properties:
      cost_rub:
        type: integer
      end_date:
        type: string
      service_name:
        type: string
      start_date:
        type: string
      user_id:
        type: string
    required:
    - cost_rub
    - service_name
    - start_date
    - user_id
    type: object
host: localhost:8080
info:
  contact: {}
  description: REST API for managing user subscriptions.
  title: Subscription Aggregator API
  version: "1.0"
paths:
  /subscriptions:
    get:
      description: Get all subscriptions
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Subscription'
            type: array
      summary: List all subscriptions
      tags:
      - subscriptions
    post:
      consumes:
      - application/json
      description: Create a new subscription
      parameters:
      - description: Subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.CreateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Create subscription
      tags:
      - subscriptions
  /subscriptions/{id}:
    delete:
      description: Delete subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
      summary: Delete subscription
      tags:
      - subscriptions
    get:
      description: Get subscription by UUID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Get subscription by ID
      tags:
      - subscriptions
    put:
      consumes:
      - application/json
      description: Update subscription by ID
      parameters:
      - description: Subscription ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated subscription data
        in: body
        name: subscription
        required: true
        schema:
          $ref: '#/definitions/model.UpdateSubscriptionRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Subscription'
        "400":
          description: Bad Request
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties: true
            type: object
      summary: Update subscription
      tags:
      - subscriptions
  /subscriptions/total-cost:
    get:
      description: Calculate total subscription cost for period with optional filters
      parameters:
      - description: User UUID
        in: query
        name: user_id
        type: string
      - description: Service name
        in: query
        name: service_name
        type: string
      - description: Start month (1-12)
        in: query
        name: start_month
        required: true
        type: integer
      - description: Start year
        in: query
        name: start_year
        required: true
        type: integer
      - description: End month (1-12)
        in: query
        name: end_month
        required: true
        type: integer
      - description: End year
        in: query
        name: end_year
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: integer
            type: object
      summary: Get total cost
      tags:
      - analytics
swagger: "2.0"